Q2/requirements.txt
streamlit
langchain
langchain-community
langchain-openai
python-dotenv
sqlite3 